+++ asdf_astropy/converters/table/__init__.py
@@ -2,6 +2,7 @@ __all__ = [
     "ColumnConverter",
     "AstropyTableConverter",
     "AsdfTableConverter",
+    "NdarrayMixinConverter",
 ]
 
-from .table import AsdfTableConverter, AstropyTableConverter, ColumnConverter
+from .table import AsdfTableConverter, AstropyTableConverter, ColumnConverter, NdarrayMixinConverter